fix: Fix test and Puller/Pusher
